Revert "Added admin layout"
This reverts commit 9463521b49c2e08243c97cfd9191d68ee8e447d0.
