+++ src/e-commerce/src/App.js
@@ -19,6 +19,11 @@ const App = () => {
   return (
     <div className="App">
       <DemoLayout/>
+      
+      <DemoCards/>
+      <DemoCheckout/>
+      <DemoProduct/>
+      
     </div>
   );
 };
