+++ src/e-commerce/src/components/layout-customer/blocker/blocker.css
@@ -2,7 +2,7 @@
     width: 100%;
     height: 100%;
     position: fixed;
-    z-index: 200;
+    z-index: 100;
     left: 0;
     top: 0;
     background-color: rgba(0, 0, 0, 0.5);
