+++ src/e-commerce/src/components/layout-customer/layout-customer.css
@@ -1,0 +1,3 @@
+.LayoutWindow{
+    padding: 75px;
+}
