+++ src/e-commerce/src/components/layout-customer/nav-bar/nav-bar.css
@@ -9,7 +9,7 @@
     align-items: center;
     padding: 0 10px;
     box-sizing: border-box;
-    z-index: 100;
+    z-index: 2000;
 }
 
 .Navbar nav {
@@ -25,28 +25,6 @@
     padding-top: 15px;
     padding-left: 20px;
 }
-
-.LogoSpaceAdmin{
-    height: 40px;
-    padding-left: 10px;
-}
-
-.WelcomeSpaceAdmin{
-    padding-top: 10px;
-}
-
-.AdminTextUnderLogo{
-    font-family: "Roboto", sans-serif;
-    font-style: italic;
-    font-size: 25px;
-    color: #f98f3e;
-    width: 100%;
-    box-sizing: border-box;
-    display: block;
-    text-align: right;
-    padding-right: 25px;
-}
-
 
 @media(max-width: 767px) {
     .Search {
@@ -73,13 +51,5 @@
 @media(max-width: 350px) {
     .CartSpace{
         display: none;
-    }
-}
-
-@media(max-width: 767px) {
-    .LogoSpaceAdmin{
-        height: 40px;
-        margin-left: auto;
-        margin-right: auto;
     }
 }
