+++ src/e-commerce/src/components/layout-customer/welcome-message/welcome.css
@@ -1,6 +1,6 @@
 .WelcomeMessage{
     font-family: "Roboto", sans-serif;
-    font-size: 21px;
+    font-size: 18px;
     color: white;
     width: 100%;
     box-sizing: border-box;
