+++ src/e-commerce/src/components/layout-customer/welcome-message/welcome.js
@@ -8,7 +8,7 @@ const WelcomeMessage = (props) => {
      }
 
     return(
-        (props.accessLevel.role == 'registeredUser' ||  props.accessLevel.role == 'admin')
+        props.accessLevel.role == 'registeredUser' 
         ? <a href={"/"} class={attachedClass}>Welcome {props.accessLevel.name}! </a> 
         : <a href={"/"} class={attachedClass}> Welcome, please log in! </a> 
     );
