--- src/e-commerce/src/components/layout/layout-admin.css
@@ -1,12 +1,0 @@
-.LayoutWindowAdmin{
-    padding-top: 140px;
-    padding-left: auto;
-    padding-right: auto;
-}
-
-
-@media(min-width: 768px) {
-    .LayoutWindowAdmin{
-        padding-left: 150px;
-    }
-}
