--- src/e-commerce/src/components/layout/layout-customer.css
@@ -1,12 +1,0 @@
-.LayoutWindowCustomer{
-    padding-top: 140px;
-    padding-left: auto;
-    padding-right: auto;
-}
-
-
-@media(min-width: 768px) {
-    .LayoutWindowCustomer{
-        padding-left: 150px;
-    }
-}
